+++ packages/elements/src/internals/machines/sign-up/router.machine.ts
@@ -4,13 +4,13 @@ import type { NonReducibleUnknown } from 'xstate';
 import { and, assign, enqueueActions, log, not, or, sendTo, setup, spawnChild, stopChild } from 'xstate';
 
 import {
+  ERROR_CODES,
   SEARCH_PARAMS,
   SIGN_IN_DEFAULT_BASE_PATH,
   SIGN_UP_DEFAULT_BASE_PATH,
   SSO_CALLBACK_PATH_ROUTE,
 } from '~/internals/constants';
-import type { ClerkElementsError } from '~/internals/errors';
-import { ClerkElementsRuntimeError } from '~/internals/errors';
+import { ClerkElementsError, ClerkElementsRuntimeError } from '~/internals/errors';
 import { ThirdPartyMachine, ThirdPartyMachineId } from '~/internals/machines/third-party';
 import { shouldUseVirtualRouting } from '~/internals/machines/utils/next';
 
@@ -39,6 +39,7 @@ export const SignUpRouterMachine = setup({
     thirdParty: ThirdPartyMachine,
   },
   actions: {
+    clearFormErrors: sendTo(({ context }) => context.formRef, { type: 'ERRORS.CLEAR' }),
     logUnknownError: snapshot => console.error('Unknown error:', snapshot),
     navigateInternal: ({ context }, { path, force = false }: { path: string; force?: boolean }) => {
       if (!context.router) return;
@@ -66,7 +67,36 @@ export const SignUpRouterMachine = setup({
         return new ClerkElementsRuntimeError('Unknown error');
       },
     }),
-    resetError: assign({ error: undefined }),
+    setFormOAuthErrors: ({ context }) => {
+      const errorOrig = context.clerk.client.signIn.firstFactorVerification.error;
+
+      if (!errorOrig) {
+        return;
+      }
+
+      let error: ClerkElementsError;
+
+      switch (errorOrig.code) {
+        case ERROR_CODES.NOT_ALLOWED_TO_SIGN_UP:
+        case ERROR_CODES.OAUTH_ACCESS_DENIED:
+        case ERROR_CODES.NOT_ALLOWED_ACCESS:
+        case ERROR_CODES.SAML_USER_ATTRIBUTE_MISSING:
+        case ERROR_CODES.OAUTH_EMAIL_DOMAIN_RESERVED_BY_SAML:
+        case ERROR_CODES.USER_LOCKED:
+          error = new ClerkElementsError(errorOrig.code, errorOrig.longMessage!);
+          break;
+        default:
+          error = new ClerkElementsError(
+            'unable_to_complete',
+            'Unable to complete action at this time. If the problem persists please contact support.',
+          );
+      }
+
+      context.formRef.send({
+        type: 'ERRORS.SET',
+        error,
+      });
+    },
     transfer: ({ context }) => context.router?.push(context.clerk.buildSignInUrl()),
   },
   guards: {
@@ -145,7 +175,7 @@ export const SignUpRouterMachine = setup({
           enqueue.spawnChild(logic, {
             id,
             systemId: id,
-            input: { basePath: context.router?.basePath, parent: self, ...input },
+            input: { basePath: context.router?.basePath, parent: self, formRef: context.formRef, ...input },
             syncSnapshot: true, // Subscribes to the spawned actor and send back snapshot events
           });
         }
@@ -176,6 +206,7 @@ export const SignUpRouterMachine = setup({
               isLoading: false,
             },
             exampleMode: event.exampleMode || false,
+            formRef: event.formRef,
           })),
           target: 'Init',
         },
@@ -189,6 +220,7 @@ export const SignUpRouterMachine = setup({
           basePath: context.router?.basePath ?? SIGN_UP_DEFAULT_BASE_PATH,
           environment: context.clerk.__unstable__environment,
           flow: 'signUp',
+          formRef: context.formRef,
           parent: self,
         }),
       }),
@@ -225,6 +257,7 @@ export const SignUpRouterMachine = setup({
     },
     Start: {
       tags: 'route:start',
+      exit: 'clearFormErrors',
       on: {
         NEXT: [
           {
@@ -329,13 +362,13 @@ export const SignUpRouterMachine = setup({
       on: {
         NEXT: {
           target: 'Start',
-          actions: 'resetError',
+          actions: 'clearFormErrors',
         },
       },
     },
     Hist: {
       type: 'history',
-      exit: 'resetError',
+      exit: 'clearFormErrors',
     },
   },
 });