+++ packages/elements/src/react/sign-in/step/step.tsx
@@ -1,4 +1,7 @@
+import { eventComponentMounted } from '@clerk/shared/telemetry';
+
 import { ClerkElementsRuntimeError } from '~/internals/errors';
+import { useTelemetry } from '~/react/utils/telemetry';
 
 import type { SignInChooseStrategyProps } from '../choose-strategy';
 import { SignInChooseStrategy, SignInForgotPassword } from '../choose-strategy';
@@ -43,6 +46,10 @@ export type SignInStepProps =
  * </SignIn.Root>
  */
 export function SignInStep(props: SignInStepProps) {
+  const telemetry = useTelemetry();
+
+  telemetry?.record(eventComponentMounted('Elements_SignInStep', { name: props.name }));
+
   switch (props.name) {
     case SIGN_IN_STEPS['start']:
       return <SignInStart {...props} />;